refactor(releases): own task route from canonical path
