+++ apps/web/app/app/(shell)/dashboard/releases/[releaseId]/tasks/page.tsx
@@ -1,75 +1,3 @@
-import { and, eq } from 'drizzle-orm';
-import { notFound, redirect } from 'next/navigation';
-import { Suspense } from 'react';
-import {
-  ReleaseTaskPage,
-  ReleaseTaskPageSkeleton,
-} from '@/components/features/dashboard/release-tasks';
-import { ReleasePlanUpgradeInterstitial } from '@/components/features/dashboard/tasks/TasksUpgradeInterstitial';
-import { APP_ROUTES } from '@/constants/routes';
-import { getCurrentUserProfile } from '@/lib/auth/session';
-import { db } from '@/lib/db';
-import { discogReleases } from '@/lib/db/schema/content';
-import { getCurrentUserEntitlements } from '@/lib/entitlements/server';
+import { ReleaseTasksRoute } from '@/app/app/(shell)/releases/[releaseId]/tasks/ReleaseTasksRoute';
 
-interface TasksPageProps {
-  readonly params: Promise<{ releaseId: string }>;
-}
-
-export default async function TasksPage({ params }: TasksPageProps) {
-  const { releaseId } = await params;
-
-  return (
-    <Suspense fallback={<ReleaseTaskPageSkeleton />}>
-      <TasksContent releaseId={releaseId} />
-    </Suspense>
-  );
-}
-
-async function TasksContent({ releaseId }: Readonly<{ releaseId: string }>) {
-  const profilePromise = getCurrentUserProfile();
-  const entitlementsPromise = getCurrentUserEntitlements();
-  const [profile, entitlements] = await Promise.all([
-    profilePromise,
-    entitlementsPromise,
-  ]);
-  const profileId = profile?.id;
-
-  if (!profileId || !profile.onboardingCompletedAt) {
-    redirect(APP_ROUTES.START);
-  }
-
-  const [release] = await db
-    .select({
-      title: discogReleases.title,
-      releaseDate: discogReleases.releaseDate,
-    })
-    .from(discogReleases)
-    .where(
-      and(
-        eq(discogReleases.id, releaseId),
-        eq(discogReleases.creatorProfileId, profileId)
-      )
-    )
-    .limit(1);
-
-  if (!release) {
-    notFound();
-  }
-
-  if (!entitlements.canAccessTasksWorkspace) {
-    return <ReleasePlanUpgradeInterstitial releaseTitle={release.title} />;
-  }
-
-  return (
-    <ReleaseTaskPage
-      profileId={profileId}
-      releaseId={releaseId}
-      releaseTitle={release.title ?? 'Release'}
-      releaseDate={release.releaseDate}
-      showMetadataAgentPanel={
-        entitlements.isAdmin || entitlements.canAccessMetadataSubmissionAgent
-      }
-    />
-  );
-}
+export default ReleaseTasksRoute;
